Add docs for typelets.
The new docs reflect parts of the README, and include a code reference
for all the types. This is set up for hosting on readthedocs, and can be
referenced using intersphinx.

Testing Done:
Built the docs. Checked for any bad links or errors.

Reviewed at https://reviews.reviewboard.org/r/14039/